+++ frontend/components/RegionDropdown.tsx
@@ -2,6 +2,7 @@ import { Region } from "../models";
 import * as React from "react";
 import { string_score } from "../string_score";
 import { classNames } from "../helpers";
+import { Dropdown } from "./Dropdown";
 
 type Props = {
   regions: Array<Region>;
@@ -16,20 +17,22 @@ export function RegionDropdown({ id, regions, selected, onSelect }: Props) {
   const [show, setShow] = React.useState<boolean>(false);
 
   let filterRef = React.useRef<HTMLInputElement>(null);
-  let containerRef = React.useRef<HTMLDivElement>(null);
-  let buttonRef = React.useRef<HTMLButtonElement>(null);
 
-  function doBlur() {
-    setShow(false);
-    if (query !== "") {
-      setQuery("");
-      setFocused(-1);
+  function onToggle(show: boolean) {
+    if (show === false) {
+      setShow(false);
+      if (query !== "") {
+        setQuery("");
+        setFocused(-1);
+      }
+    } else {
+      setShow(true);
     }
   }
 
   function doChange(region: Region) {
     onSelect(region);
-    doBlur();
+    onToggle(false);
   }
 
   function updateQuery(query: string) {
@@ -43,26 +46,6 @@ export function RegionDropdown({ id, regions, selected, onSelect }: Props) {
       $filter.focus();
     }
   }, [show]);
-
-  React.useEffect(() => {
-    const handler = (event: MouseEvent) => {
-      let element = containerRef.current;
-      let clickedInside =
-        element &&
-        event.target instanceof Node &&
-        element.contains(event.target);
-
-      if (!clickedInside) {
-        doBlur();
-      }
-    };
-
-    document.addEventListener("click", handler);
-
-    return () => {
-      document.removeEventListener("click", handler);
-    };
-  });
 
   // make a copy of the regions and add scores
   let list = React.useMemo<Array<{ region: Region; score: number }>>(
@@ -126,53 +109,40 @@ export function RegionDropdown({ id, regions, selected, onSelect }: Props) {
     } else if (evt.key === "ArrowDown") {
       setFocused(Math.min(focused + 1, list.length - 1));
     } else if (evt.key === "Escape") {
-      doBlur();
+      onToggle(false);
     }
   }
 
+  let button = (
+    <>
+      <span className="dropdown-label">{selected?.name ?? <>...</>}</span>{" "}
+      <span className="caret"></span>
+    </>
+  );
+
   return (
-    <div className={classNames({ show, dropdown: true })} ref={containerRef}>
-      <button
-        className="classbtn btn-primary dropdown-toggle medium dropdown-caret-styling region-dropdown"
-        type="button"
-        onClick={() => setShow(!show)}
-        ref={buttonRef}
-        aria-haspopup="true"
-        aria-expanded={show}
-        id={id}
-      >
-        <span className="region-dropdown-label">
-          {selected?.name ?? <>...</>}
-        </span>
-        <span className="caret"></span>
-      </button>
-      <div
-        className={classNames("dropdown-menu", { show })}
-        aria-labelledby={id}
-        x-placement="bottom-start"
-      >
-        <div className="px-3 py-2">
-          <input
-            ref={filterRef}
-            className="form-control"
-            onKeyDown={handleKeyDown}
-            onChange={(evt) => updateQuery(evt.target.value)}
-            value={query}
-            type="text"
-            placeholder="Filter..."
-          />
-        </div>
-        <div className="dropdown-divider"></div>
-        <div
-          className="dropdown-list"
-          style={{
-            overflowY: "auto",
-            maxHeight: "50vh",
-          }}
-        >
-          {dropdownEntries}
-        </div>
+    <Dropdown show={show} id={id} button={button} onToggle={onToggle}>
+      <div className="px-3 py-2">
+        <input
+          ref={filterRef}
+          className="form-control"
+          onKeyDown={handleKeyDown}
+          onChange={(evt) => updateQuery(evt.target.value)}
+          value={query}
+          type="text"
+          placeholder="Filter..."
+        />
       </div>
-    </div>
+      <div className="dropdown-divider"></div>
+      <div
+        className="dropdown-list"
+        style={{
+          overflowY: "auto",
+          maxHeight: "50vh",
+        }}
+      >
+        {dropdownEntries}
+      </div>
+    </Dropdown>
   );
 }
